requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
clickhouse-connect==0.8.18
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lz4==4.4.4
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
zstandard==0.23.0
